+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultSampleService.java
@@ -9,7 +9,6 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
@@ -936,13 +935,6 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
   @Override
   public Collection<String> listSampleTypes() throws IOException {
     return sampleStore.listSampleTypes();
-  }
-
-  @Override
-  public Map<String, Integer> getSampleColumnSizes() throws IOException {
-    return ValidationUtils.adjustNameLength(
-        ValidationUtils.adjustLength(sampleStore.getSampleColumnSizes(), "alias", namingScheme.sampleAliasLengthAdjustment()),
-        namingScheme);
   }
 
   @Override