+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ Changes:
 
 * Added single Edit Dilution page
 * Changed Edit Library page to show validation messages instead of error page when saving fails
+* Changed Edit Pool page to show validation messages instead of error page when saving fails
 * Fixed missing Notes section on Edit Sample page
 
 # 0.2.174